% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dictionary}
\alias{base_dictionary}
\title{Base dictionary with unigrams}
\format{
A data.frame with four columns: to_search, to_replace, points, id.
}
\usage{
data(base_dictionary)
}
\description{
Data contains English unigrams, with their replacements, points and ids.
}
\examples{
data(base_dictionary)
}
\references{
Created with inspiration from Kevin Atkinson (2020) (\href{http://wordlist.aspell.net}{SCOWL and Friends})
}
\keyword{datasets}
